+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- CertReq:
+  - Fix error when ProviderName parameter is not encapsulated in
+    double quotes - fixes [Issue #185](https://github.com/PowerShell/CertificateDsc/issues/185).
+
 ## 4.6.0.0
 
 - CertReq: